Make it work on stable Rust
Due to `cfg(test)` being set, this library is not able to be used outside
of tests. Importing it would fail with:

    error[E0432]: unresolved import `peekbufread::PeekRead`

With the `features(test)` being set it's not possible to use it
on stable Rust. Importing it would fail with:

    error[E0554]: `#![feature]` may not be used on the stable release channel

With separating things into the usual directory structure of the library
source, tests and benchmarks, it's now possible to use this library also
on stable Rust.

For the benchmarks nightly is still required.